룸 기능 2차 수정
